fix: fix wranings

+++ src/0005.cpp
@@ -9,13 +9,13 @@ using namespace std;
 namespace leetcode_0005 {
 string key = "0005";
 class Solution {
-public:
+ public:
   string longestPalindrome(string s) {
-    int length = s.length();
+    size_t length = s.length();
     string result = "";
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
       // single
-      int max_target_step = min(i, length - 1 - i);
+      size_t max_target_step = min(i, length - 1 - i);
 
       longestIntermediatePalindrome(s, length, &result, max_target_step, i, i);
 
@@ -24,23 +24,22 @@ public:
         return result;
       }
       max_target_step = min(i, length - 1 - (i + 1));
-      longestIntermediatePalindrome(s, length, &result, max_target_step, i,
-                                    i + 1);
+      longestIntermediatePalindrome(s, length, &result, max_target_step, i, i + 1);
     }
     return result;
   }
 
-private:
-  void longestIntermediatePalindrome(string s, int length, string *result,
-                                     int max_target_step, int m, int n) {
-    int max_length = max_target_step * 2 + n - m + 1;
+ private:
+  void longestIntermediatePalindrome(string s, int length, string *result, unsigned int max_target_step, unsigned int m,
+                                     unsigned int n) {
+    unsigned int max_length = max_target_step * 2 + n - m + 1;
     if ((*result).length() > max_length) {
       return;
     }
-    int step = 0;
+    unsigned int step = 0;
     while (step <= max_target_step) {
       if (s.at(m - step) != s.at(n + step)) {
-        int tmp = (step - 1) * 2 + n - m + 1;
+        unsigned int tmp = (step - 1) * 2 + n - m + 1;
         if (tmp > (*result).length()) {
           *result = s.substr(m - step + 1, tmp);
         }
@@ -58,9 +57,9 @@ int main() {
   strs.push_back("babad");
   strs.push_back("cbbd");
   strs.push_back("a");
-  for (int i = 0; i < strs.size(); i++) {
+  for (size_t i = 0; i < strs.size(); i++) {
     cout << Solution().longestPalindrome(strs.at(i)) << endl;
   }
   return 0;
 }
-} // namespace leetcode_0005
+}  // namespace leetcode_0005
