+++ src/0010.cpp
@@ -7,8 +7,10 @@ using namespace std;
 namespace leetcode_0010 {
 string key = "0010";
 class Solution {
-public:
-  bool isMatch(string s, string p) { return false; }
+ public:
+  bool isMatch(string s, string p) {
+    return false;
+  }
 };
 int main() {
   std::cout << key << ":正则表达式匹配" << std::endl;
@@ -35,11 +37,9 @@ int main() {
   p.push_back("mis*is*p*.");
   m.push_back(false);
 
-  for (int i = 0; i < s.size(); i++) {
-    cout << (Solution().isMatch(s.at(i), p.at(i)) == m.at(i) ? "pass"
-                                                             : "falied")
-         << endl;
+  for (size_t i = 0; i < s.size(); i++) {
+    cout << (Solution().isMatch(s.at(i), p.at(i)) == m.at(i) ? "pass" : "falied") << endl;
   }
   return 0;
 }
-} // namespace leetcode_0010
+}  // namespace leetcode_0010
